requirements.txt
langchain
langchain-openai
web3
python-dotenv
requests
